+++ ricaun.Revit.Github.Example/Revit/Commands/Command.cs
@@ -17,60 +17,9 @@ namespace ricaun.Revit.Github.Example.Revit.Commands
         {
             UIApplication uiapp = commandData.Application;
 
-            var request = new GithubRequestService("ricaun-io", "ricaun.Nuke.PackageBuilder");
+            var request = new GithubRequestService("ricaun-io", "ricaun.Revit.Github");
 
-            request.Initialize();
-
-            //var text = request.DownloadLast();
-            //System.Windows.MessageBox.Show($">>> {text}");
-
-            //var githubBundleService = new Services.GithubBundleService("ricaun-io", "ricaun.Nuke.PackageBuilder");
-
-            //foreach (var item in githubBundleService.GetBundleModels(Assembly.GetExecutingAssembly()))
-            //{
-            //    Console.WriteLine(item);
-            //}
-
-
-
-            //var a = Assembly.GetExecutingAssembly();
-            //var version = a.GetName().Version.ToString();
-
-
-
-            var githubService = new Github.Services.GithubBundleService("ricaun-io", "ricaun.Nuke.PackageBuilder2");
-            var task = Task.Run(async () =>
-                {
-                    try
-                    {
-                        await Task.Delay(100);
-                        var text = await githubService.DownloadStringAsync("https://api.github.com/repos/ricaun-io/ricaun.Nuke.PackageBuilder/releases/latest");
-                        Console.WriteLine($"Result {text}");
-
-
-                        var bundleModels = await githubService.GetBundleModelsAsync();
-                        Console.WriteLine(bundleModels.Count());
-                        foreach (var bundleModel in bundleModels)
-                        {
-                            Console.WriteLine(bundleModel);
-                        }
-
-                    }
-                    catch (Exception ex)
-                    {
-                        Console.WriteLine(ex);
-                    }
-                });
-
-
-            //TaskDialog.Show("Hello", "Luiz");
-
-            //task.GetAwaiter().GetResult();
-
-            //var service = new Services.GithubBundleService("ricaun-io", "ricaun.Nuke.PackageBuilder");
-            //Console.WriteLine($"1.2.4 - {version} {service.GetBundleDownloadUrl("1.2.4", version)}");
-            //Console.WriteLine($"0.0.10 - {version} {service.GetBundleDownloadUrl("0.0.10", version)}");
-            //Console.WriteLine($"0.0.10 {service.GetBundleDownloadUrl("0.0.10")}");
+            request.Initialize((text) => { Console.WriteLine(text); });
 
             return Result.Succeeded;
         }