+++ CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] / 2022-05-21
+## [0.2.0] / 2023-01-31
+### Features
+- [x] Initialize Auto Download Last Version - `Console log`
+- [x] Update Example Project to Update Last Version
+
+## [0.1.0] / 2022-05-21
 ### Features
 - [x] Download Bundle From Github
 - [ ] Improve Library Files
@@ -29,4 +34,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add Example
 
 [vNext]: ../../compare/1.0.0...HEAD
+[0.2.0]: ../../compare/0.1.0...0.2.0
+[0.1.0]: ../../compare/0.1.0
 [1.0.0]: ../../compare/1.0.0